% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sawhet-package.R
\docType{package}
\name{sawhet-package}
\alias{sawhet}
\alias{sawhet-package}
\title{sawhet: STEM Automated Writing Help Tool}
\description{
An online web form to collect student lab reports in a structured format.
}
\author{
\strong{Maintainer}: Evan Anway \email{evan.anway@gmail.com}

}
\keyword{internal}
